Update Moment start of week in i18n.js instead
